Re-orgainze TestHarness tests into individual files so they can run in parallel
(refs #8700)
