+++ python/chigger/tests/contours/tests
@@ -18,6 +18,7 @@
     type = RunException
     command = block_elem.py
     expect_err = "ContourFilter currently only works with nodal variables."
+    display_required = True
   [../]
 
   [./levels]
