+++ python/mooseutils/tests/tests
@@ -8,6 +8,7 @@
   [./mooseMessageDialog]
     type = PythonUnitTest
     input = test_mooseMessageDialog.py
+    display_required = True
   [../]
 
   [./moose_data_frame]
